fix!: detach format lifetime from worksheet
At the moment the lifetime of `Format` inside
`IntoExcelData::write_to_format` is bound to the lifetime of the
worksheet. However, this is not actually required anywhere in the code,
and with this constraint it is not possible to create abstractions with
an internal formatter because, once the ownership of the data is taken
by `write_with_format`, the stored format cannot outlive the call.

This breaking change should simplify the management of formatters.